Refactor: rename Script to a mire appropriate ScriptController name
