+++ Makefile.am
@@ -150,8 +150,8 @@ nzbget_SOURCES = \
 	daemon/util/Container.h \
 	daemon/util/Observer.cpp \
 	daemon/util/Observer.h \
-	daemon/util/Script.cpp \
-	daemon/util/Script.h \
+	daemon/util/ScriptController.cpp \
+	daemon/util/ScriptController.h \
 	daemon/util/Thread.cpp \
 	daemon/util/Thread.h \
 	daemon/util/Service.cpp \
