+++ daemon/main/Maintenance.h
@@ -22,7 +22,7 @@
 
 #include "NString.h"
 #include "Thread.h"
-#include "Script.h"
+#include "ScriptController.h"
 #include "Log.h"
 #include "Util.h"
