+++ daemon/postprocess/Cleanup.h
@@ -25,7 +25,7 @@
 #include "Log.h"
 #include "Thread.h"
 #include "DownloadInfo.h"
-#include "Script.h"
+#include "ScriptController.h"
 
 class MoveController : public Thread, public ScriptController
 {
