+++ nzbget.vcxproj
@@ -273,7 +273,7 @@
     <ClCompile Include="daemon\remote\XmlRpc.cpp" />
     <ClCompile Include="daemon\util\Log.cpp" />
     <ClCompile Include="daemon\util\Observer.cpp" />
-    <ClCompile Include="daemon\util\Script.cpp" />
+    <ClCompile Include="daemon\util\ScriptController.cpp" />
     <ClCompile Include="daemon\util\Service.cpp" />
     <ClCompile Include="daemon\util\Thread.cpp" />
     <ClCompile Include="daemon\util\NString.cpp" />
@@ -392,7 +392,7 @@
     <ClInclude Include="daemon\remote\XmlRpc.h" />
     <ClInclude Include="daemon\util\Log.h" />
     <ClInclude Include="daemon\util\Observer.h" />
-    <ClInclude Include="daemon\util\Script.h" />
+    <ClInclude Include="daemon\util\ScriptController.h" />
     <ClInclude Include="daemon\util\Service.h" />
     <ClInclude Include="daemon\util\Thread.h" />
     <ClInclude Include="daemon\util\NString.h" />
